Move timestamp helper to sdk
