+++ sdk/src/lib.rs
@@ -45,6 +45,7 @@ pub mod short_vec;
 pub mod slot_hashes;
 pub mod slot_history;
 pub mod stake_history;
+pub mod stake_weighted_timestamp;
 pub mod system_instruction;
 pub mod system_program;
 pub mod sysvar;
